use u32 as BlockNumber in mocks

+++ precompiles/author-mapping/src/mock.rs
@@ -36,7 +36,7 @@ use sp_runtime::{
 
 pub type AccountId = Account;
 pub type Balance = u128;
-pub type BlockNumber = u64;
+pub type BlockNumber = u32;
 
 type UncheckedExtrinsic = frame_system::mocking::MockUncheckedExtrinsic<Runtime>;
 type Block = frame_system::mocking::MockBlock<Runtime>;
